requirements.txt
# Paquetes principales
inference==0.33.0 
inference-cli==0.33.0
supervision==0.25.1
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
requests>=2.31.0
aiohttp>=3.9.1
pydantic-settings>=2.0.0

# Herramientas de desarrollo (opcional)
black>=24.1.1               # Para formateo de código
isort>=5.13.2               # Para ordenar imports
